+++ authentication-server/src/main/java/com/wesell/authenticationserver/controller/AdminAuthController.java
@@ -2,8 +2,8 @@ package com.wesell.authenticationserver.controller;
 
 import com.wesell.authenticationserver.controller.dto.request.AdminAuthIsForcedRequestDto;
 import com.wesell.authenticationserver.controller.dto.request.AdminAuthRoleRequestDto;
+import com.wesell.authenticationserver.service.AuthUserService;
 import com.wesell.authenticationserver.service.dto.response.AdminAuthIsForcedResponseDto;
-import com.wesell.authenticationserver.service.AdminAuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,12 +13,13 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("admin-auth-server")
 @RequiredArgsConstructor
 public class AdminAuthController {
-    private final AdminAuthService adminAuthService;
+
+    private final AuthUserService authUserService;
 
     @PutMapping("change-role/{uuid}")
     public ResponseEntity<String> changeUserRole(@PathVariable String uuid, @RequestBody AdminAuthRoleRequestDto adminAuthRoleRequestDto) {
         try {
-            adminAuthService.updateRole(uuid, adminAuthRoleRequestDto.getRole());
+            authUserService.updateRole(uuid, adminAuthRoleRequestDto.getRole());
             return new ResponseEntity<>("User role changed successfully", HttpStatus.OK);
         } catch (Exception e) {
             return new ResponseEntity<>("Failed to change user role: " + e.getMessage(), HttpStatus.BAD_REQUEST);
@@ -27,7 +28,7 @@ public class AdminAuthController {
 
     @PatchMapping("updateIsForced")
     public ResponseEntity<AdminAuthIsForcedResponseDto> updateIsForced(@RequestBody AdminAuthIsForcedRequestDto requestDto) {
-        AdminAuthIsForcedResponseDto responseDto = adminAuthService.updateIsForced(requestDto);
+        AdminAuthIsForcedResponseDto responseDto = authUserService.updateIsForced(requestDto);
         return new ResponseEntity<>(responseDto, HttpStatus.OK);
     }
 }
