--- authentication-server/src/main/java/com/wesell/authenticationserver/domain/repository/AdminAuthUserRepository.java
@@ -1,8 +1,0 @@
-package com.wesell.authenticationserver.domain.repository;
-
-import com.wesell.authenticationserver.domain.entity.AuthUser;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AdminAuthUserRepository extends JpaRepository<AuthUser, Long> {
-    AuthUser findByUuid(String uuid);
-}
